# registry-api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "StarForge Registry API",
    "version": "1.0.0",
    "license": { "name": "MIT", "url": "https://opensource.org/license/mit/" },
    "description": "HTTP API for the StarForge template registry. Authenticated operations use a JWT access token in the Authorization header as `Bearer <token>`."
  },
  "servers": [
    {
      "url": "http://{host}:{port}",
      "description": "Registry API server",
      "variables": {
        "host": { "default": "localhost" },
        "port": { "default": "3000" }
      }
    }
  ],
...
  "paths": {
...
    "/api/auth/signup": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Create an account",
        "operationId": "signUp",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SignupRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created and access token issued",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Log in",
        "operationId": "login",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Access token issued",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/auth/verify": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Verify an access token",
        "operationId": "verifyToken",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Token is valid",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenVerification" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/api/reviews/template/{templateId}": {
      "get": {
        "tags": ["Reviews"],
        "summary": "List reviews for a template",
        "operationId": "getTemplateReviews",
        "security": [{ "bearerAuth": [] }, {}],
        "parameters": [
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Reviews and total count",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ReviewsResponse" }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/reviews/template/{templateId}/reviews": {
      "post": {
        "tags": ["Reviews"],
        "summary": "Create or update a template review",
        "operationId": "upsertTemplateReview",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ReviewRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Existing review updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          },
          "201": {
            "description": "Review created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT issued by POST /api/auth/signup or POST /api/auth/login. Send as `Authorization: Bearer <token>`."
      }
    },
...
    "responses": {
      "BadRequest": {
        "description": "Invalid or incomplete request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing, invalid, or expired Bearer token, or invalid credentials",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
      "Conflict": {
        "description": "Account or template version conflicts with an existing record",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "InternalError": {
        "description": "The request could not be completed",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": { "error": { "type": "string" } }
      },
...
      "SignupRequest": {
        "type": "object",
        "required": ["email", "username", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "username": { "type": "string" },
          "password": { "type": "string", "minLength": 8, "format": "password" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "format": "password" }
        }
      },
      "AuthResponse": {
        "type": "object",
        "required": ["success", "message", "token", "username"],
        "properties": {
          "success": { "type": "boolean", "enum": [true] },
          "message": { "type": "string" },
          "token": { "type": "string", "description": "JWT Bearer access token." },
          "username": { "type": "string" }
        }
      },
      "TokenVerification": {
        "type": "object",
        "required": ["success", "user"],
        "properties": {
          "success": { "type": "boolean", "enum": [true] },
          "user": { "type": "object", "additionalProperties": true }
        }
      },
...
      "ReviewRequest": {
        "type": "object",
        "required": ["rating"],
        "properties": {
          "rating": { "type": "integer", "minimum": 1, "maximum": 5 },
          "comment": { "type": "string" }
        }
      },
      "Review": {
        "type": "object",
        "required": ["id", "rating", "created_at"],
        "properties": {
          "id": { "type": "string" },
          "rating": { "type": "integer" },
          "comment": { "type": "string", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "ReviewsResponse": {
        "type": "object",
        "required": ["success", "reviews", "total"],
        "properties": {
          "success": { "type": "boolean" },
          "reviews": { "type": "array", "items": { "$ref": "#/components/schemas/Review" } },
          "total": { "type": "integer" }
        }
      },
      "MessageResponse": {
        "type": "object",
        "required": ["success", "message"],
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": "string" }
        }
      }
    }
  }
}